backend/requirements.txt
flask
flask-cors
openai
python-dotenv
pymongo
flask-pymongo
PyPDF2
PyMuPDF
PyJWT
bcrypt
